requirements.txt
chroma-hnswlib==0.7.3
chromadb==0.4.20
fastapi==0.105.0
fastapi-cli==0.0.4
langchain==0.3.13
langchain-core==0.3.28
langsmith==0.1.125
mypy-extensions==1.0.0
numpy==1.24.4
openai==1.58.1
Pillow==10.1.0
posthog==3.1.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core
Pygments==2.17.2
pymongo==3.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
jinja2
tiktoken==0.8.0
motor==2.5.1
whisper==1.1.10
pyttsx3==2.98
sounddevice==0.5.1
scipy==1.10.1
openai-whisper==20240930
python-multipart==0.0.17
ffmpeg-python==0.2.0
langdetect==1.0.9
langchain-openai==0.2.14